feat:adding prerendering images in gallery section
